+++ src/panel.ts
@@ -12,53 +12,6 @@ export class Panel {
     this._settings = settings;
   }
 
-  private getPanelTitleHtml(): string {
-    let panelTitleHtml: string = '';
-
-    if (this._settings.showPanelTitle) {
-      panelTitleHtml = `
-        <div id="panel-title" style="height:${this._settings.lineHeight}px;"
-          ondragover="dragOverTitle(event);" ondragleave="dragLeave(event);" ondrop="dropOnTitle(event);" ondragend="dragLeave(event);">
-          <span class="fas fa-star" style="color:${this._settings.foreground};"></span>
-          <span class="title" style="color:${this._settings.foreground};">FAVORITES</span>
-        </div>
-      `;
-    }
-    return panelTitleHtml;
-  }
-
-  // create HTML for each favorite
-  private getFavoritesHtml(): string {
-    const favsHtml: any = [];
-
-    for (const favorite of this._favs.all) {
-      const foreground = this._settings.foreground;
-      const background = this._settings.background;
-      const hoverBg = this._settings.hoverBackground;
-      const dividerColor = this._settings.dividerColor;
-
-      let typeIconHtml: string = '';
-      if (this._settings.showTypeIcons) {
-        typeIconHtml = `<span class="fas ${FavoriteDesc[favorite.type].icon}" style="color:${foreground};"></span>`;
-      }
-
-      favsHtml.push(`
-        <div id="favorite" data-id="${favorite.value}" draggable="${this._settings.enableDragAndDrop}"
-          onclick="favsClick(event);" oncontextmenu="favsContext(event);" onMouseOver="this.style.background='${hoverBg}';" onMouseOut="this.style.background='none';"
-          ondragstart="dragStart(event);" ondragover="dragOver(event, '${hoverBg}');" ondragleave="dragLeave(event);" ondrop="drop(event);" ondragend="dragEnd(event);"
-          style="height:${this._settings.lineHeight}px;min-width:${this._settings.minFavWidth}px;max-width:${this._settings.maxFavWidth}px;background:${background};border-color:${dividerColor};color:${foreground};">
-          <span class="favorite-inner" style="border-color:${dividerColor};">
-            ${typeIconHtml}
-            <span class="title" title="${favorite.title}">
-              ${favorite.title}
-            </span>
-          </span>
-        </div>
-      `);
-    }
-    return favsHtml.join('\n');
-  }
-
   /**
    * Register plugin panel and update webview for the first time.
    */
@@ -69,10 +22,10 @@ export class Panel {
     await joplin.views.panels.addScript(this._panel, './webview.js');
     await joplin.views.panels.onMessage(this._panel, async (message: any) => {
       if (message.name === 'favsAddFolder') {
-        await joplin.commands.execute('favsAddFolder', message.id);
+        await joplin.commands.execute('favsAddFolder', message.id, message.targetId);
       }
       if (message.name === 'favsAddNote') {
-        await joplin.commands.execute('favsAddNote', message.id);
+        await joplin.commands.execute('favsAddNote', message.id, message.targetId);
       }
       if (message.name === 'favsEdit') {
         await joplin.commands.execute('favsEditFavorite', message.id);
@@ -98,6 +51,56 @@ export class Panel {
     await this.updateWebview();
   }
 
+  private getPanelTitleHtml(): string {
+    let panelTitleHtml: string = '';
+
+    // ondragover="dragOver(event, '${this._settings.hoverBackground}');" ondragleave="dragLeave(event);" ondrop="drop(event);" ondragend="dragEnd(event);"
+    if (this._settings.showPanelTitle) {
+      const fg = this._settings.foreground;
+
+      panelTitleHtml = `
+        <div id="panel-title" draggable="false" style="height:${this._settings.lineHeight}px;">
+          <span class="fas fa-star" style="color:${fg};"></span>
+          <span class="title" style="color:${fg};">FAVORITES</span>
+        </div>
+      `;
+    }
+    return panelTitleHtml;
+  }
+
+  // create HTML for each favorite
+  private getFavoritesHtml(): string {
+    const favsHtml: any = [];
+
+    for (const favorite of this._favs.all) {
+      const fg = this._settings.foreground;
+      const bg = this._settings.background;
+      const hoverBg = this._settings.hoverBackground;
+      const dividerColor = this._settings.dividerColor;
+
+      let typeIconHtml: string = '';
+      if (this._settings.showTypeIcons) {
+        typeIconHtml = `<span class="fas ${FavoriteDesc[favorite.type].icon}" style="color:${fg};"></span>`;
+      }
+
+      // ondblclick="dblclick(event)"
+      favsHtml.push(`
+        <div id="favorite" data-id="${favorite.value}" data-bg="${bg}" draggable="${this._settings.enableDragAndDrop}" title="${favorite.title}"
+          onclick="favsClick(event)" oncontextmenu="context(event)" onmouseover="setBackground(event,'${hoverBg}')" onmouseout="resetBackground(this)"
+          ondragstart="dragStart(event)" ondragover="dragOver(event, '${hoverBg}')" ondragleave="dragLeave(event)" ondrop="drop(event)" ondragend="dragEnd(event)"
+          style="height:${this._settings.lineHeight}px;min-width:${this._settings.minFavWidth}px;max-width:${this._settings.maxFavWidth}px;background:${bg};border-color:${dividerColor};color:${fg};">
+          <span class="favorite-inner" style="border-color:${dividerColor};">
+            ${typeIconHtml}
+            <span class="title" title="${favorite.title}">
+              ${favorite.title}
+            </span>
+          </span>
+        </div>
+      `);
+    }
+    return favsHtml.join('\n');
+  }
+
   async updateWebview() {
     const panelTitleHtml: string = this.getPanelTitleHtml();
     const favsHtml: string = this.getFavoritesHtml();
@@ -105,8 +108,10 @@ export class Panel {
     // add entries to container and push to panel
     await joplin.views.panels.setHtml(this._panel, `
       <div id="container" style="background:${this._settings.background};font-family:'${this._settings.fontFamily}',sans-serif;font-size:${this._settings.fontSize};">
-        <div id="favs-container">
-          ${panelTitleHtml}
+        ${panelTitleHtml}
+        <div id="favs-container" draggable="${this._settings.enableDragAndDrop}"
+          ondragover="dragOver(event, '${this._settings.hoverBackground}');" ondragleave="dragLeave(event);" ondrop="drop(event);" ondragend="dragEnd(event);"
+          style="height:${this._settings.lineHeight}px;">
           ${favsHtml}
         </div>
       </div>