+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Addded
 
 - Ability to drag & drop notes from [note-tabs plugin](https://github.com/benji300/joplin-note-tabs) to add as favorite
-- Ability to rename favorites directly in panel
+- Ability to rename and delete favorites directly in panel in the vertical layout
+  - Via new hover buttons on the right side
 
 ### Changed
 
-- Drag & drop behavior to add notebooks and notes/to-dos
+- Drag & drop behavior to add notebooks, notes or to-dos
   - Move them onto the panel to add new favorite at the dropped position
 
 ## [1.1.0] - 2021-01-21