+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Drag & drop behavior to add notebooks, notes or to-dos
   - Move them onto the panel to add new favorite at the dropped position
+- Scroll horizontal without holding `Shift` key
 
 ## [1.1.0] - 2021-01-21