+++ CHANGELOG.md
@@ -7,17 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Addded
+- None
+
+## [1.2.0] - 2021-02-17
+
+### Added
 
 - Ability to drag & drop notes from [note-tabs plugin](https://github.com/benji300/joplin-note-tabs) to add as favorite
-- Ability to rename and delete favorites directly in panel in the vertical layout
+- Ability to rename and delete favorites directly in panel (vertical layout only)
   - Via new hover buttons on the right side
 
 ### Changed
 
 - Drag & drop behavior to add notebooks, notes or to-dos
   - Move them onto the panel to add new favorite at the dropped position
-- Scroll horizontal without holding `Shift` key
+- Scroll horizontally without holding `Shift` key
+
+### Fixed
+
+- Search favorites with phrases in query cannot be opened, edited or deleted ([#4](https://github.com/benji300/joplin-favorites/issues/4))
 
 ## [1.1.0] - 2021-01-21